+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/25 19:15:07 by emadriga          #+#    #+#             */
-/*   Updated: 2021/11/13 13:07:03 by emadriga         ###   ########.fr       */
+/*   Updated: 2021/11/13 20:58:55 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,4 +70,6 @@ int		ft_lstsize(t_list *lst);
 t_list	*ft_lstlast(t_list *lst);
 int		get_next_line(int fd, char **line);
 char	*ft_strreplace(const char *str, char const *oldset, char const *newset);
+char	*ft_strreplaceat(const char *str, char const *oldset, \
+		char const *newset, size_t index);
 #endif